[Perf] Better handling of telemetry benchmark failures
When a benchmark fails, the perf_results.json file is usually not
created. But the test script still tries to move the file, causing
an exception. This exception often confuses the users (e.g.,
https://crbug.com/365399707#comment2). This CL modifies the script
to avoid this exception.

Bug: 365399707
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5851647
Commit-Queue: John Chen <johnchen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1354008}

+++ testing/scripts/run_performance_tests.py
@@ -590,7 +590,15 @@ def execute_telemetry_benchmark(command_generator,
 
     if not no_output_conversion:
       expected_perf_filename = os.path.join(temp_dir, 'histograms.json')
-      shutil.move(expected_perf_filename, output_paths.perf_results)
+      if os.path.exists(expected_perf_filename):
+        shutil.move(expected_perf_filename, output_paths.perf_results)
+      elif return_code:
+        print(f'The benchmark failed with status code {return_code}, '
+              'and did not produce perf results output. '
+              'Check benchmark output for more details.')
+      else:
+        print('The benchmark returned a success status code, '
+              'but did not product perf results output.')
 
     csv_file_path = os.path.join(temp_dir, 'results.csv')
     if os.path.isfile(csv_file_path):
